[arduino] Fix warnings in arduino-cli build
Unused parameters lose their names, the shifts that build 24- and 28-bit
values are widened so they survive a 16-bit int, and the Insn state slot
becomes a byte buffer instead of a type-punned uintptr_t.

+++ src/array_memory.h
@@ -48,7 +48,7 @@ struct ArrayMemory {
             if (_endian == ENDIAN_BIG) {
                 quad = (quad << 16) | readUint16();
             } else {
-                quad |= readUint16() << 16;
+                quad |= static_cast<uint32_t>(readUint16()) << 16;
             }
             return quad;
         }
